requirements.txt
fastapi
uvicorn
groq
python-dotenv
instructor
pydantic